Reorganized code a bit. Exetended default LoRA name patterns to consume one of unsupported patterns.
